// RealFlix/ClientApp/src/app/models/show.ts
import { Genre } from './enums';

export class Show {
  Id: number;
  Url: string;
  Name: string;
  Type: string;
  Language: string;
  Genres: string;
  Status: string;
  OfficialSite: string;
  WebChannel: string;
  Image: string;
  Summary: string;
  ScheduledTime: Date;
  ScheduledDays: number;
  KeyWords: string;
}